fix(doctor): split the bootstrap probe so it is not a god file
CI's self-scan caught a real regression in the previous commit: landing the
native probe as one 1124-line file tripped the repo's own structural gate,
`god_file_count 32 -> 33` (`god_file: loc > 800 || (functions > 25 && loc >
400)`), which cost 120 quality points and failed `run execute` with exit 10
on all three platforms.

The probe is now four cohesive modules instead of one file:

  doctor_bootstrap/mod.rs     envelope assembly, missing list, CLI surface
  doctor_bootstrap/config.rs  pipeline.config.json and repo resolution
  doctor_bootstrap/paths.rs   platform derivation and path resolution
  doctor_bootstrap/probe.rs   tool presence and command-output probing

No file is a god file and behavior is unchanged; the split also let several
helpers grow real unit tests of their own (repo-alias fallback, rooted vs
relative sentruxPath, reverse-lookup edge cases, trailing-separator trimming),
so coverage went up rather than moving around.

Two follow-on corrections the split required:
- the registry entrypoint and toolchainDigestEvidence now name the four module
  files; `orchestrate --action Validate` rejected the stale single-file path,
  which is the manifest reconciliation check doing its job
- `--config`/`--platform`/`--pipeline-root` argument parsing was rewritten as a
  flat match over (token, value) instead of nested closures

Residual metric movement, recorded in .sentrux/rules.toml with evidence and
baselined: coupling_score 45.26 -> 45.79, quality_signal 3607 -> 3603. The
`[constraints]` thresholds are untouched. coupling_score is import_lines /
file_count * 10, and this tree's average is held down by large PowerShell
files carrying almost no import lines, so any idiomatic Rust module sits above
it. Closing the gap would have meant dropping ~13 `use` lines for inline fully
qualified paths — worse code in service of a line-counting metric. The real
regression, the god file, was fixed rather than baselined; god_file_count and
cycle_count are unchanged at 32 and 0.

Verification:
- code-intel run execute --repo . (the exact CI self-scan command): exit 0,
  outcome completed, every node verdict pass
- code-intel sentrux gate .: No degradation detected
- cargo test -p code-intel: 45 suites, 0 failed
- cargo fmt -p code-intel -- --check: clean
- orchestrate --action Validate: ok, 0 errors, registryAudit ok
- test-regression-fixes.ps1: 49 passed; test-doctor-repo-config-resolution.ps1: PASS

Refs #48
