+++ crates/code-intel-cli/src/doctor_adapter.rs
@@ -14,7 +14,7 @@ use crate::capability::sha256_hex;
 // integration tests pull this adapter into their own crate via `#[path]`, and
 // those roots do not declare the binary's module list. Same convention the
 // adapter already used for `tool_path`.
-#[path = "doctor_bootstrap.rs"]
+#[path = "doctor_bootstrap/mod.rs"]
 mod doctor_bootstrap;
 
 pub(crate) fn execute(
@@ -494,7 +494,10 @@ mod tests {
             .unwrap();
         for relative in [
             "crates/code-intel-cli/src/doctor_adapter.rs",
-            "crates/code-intel-cli/src/doctor_bootstrap.rs",
+            "crates/code-intel-cli/src/doctor_bootstrap/mod.rs",
+            "crates/code-intel-cli/src/doctor_bootstrap/config.rs",
+            "crates/code-intel-cli/src/doctor_bootstrap/paths.rs",
+            "crates/code-intel-cli/src/doctor_bootstrap/probe.rs",
             "crates/code-intel-cli/src/capability_inventory.rs",
         ] {
             let actual = sha256_hex(&fs::read(root.join(relative)).unwrap());
